Index trial groups in Elasticsearch
As part of https://gitlab.com/gitlab-org/gitlab/-/issues/220386 we want
to index trial groups in Elasticsearch so they can make use of Advanced
search.

This change also refactored the existing logic around indexing paid
groups a bit:

1. Removing the feature flag that has been on for a while now along
with the documentation explaining this feature flag. We already have
separate docs explaining that Advanced Global Search is enabled for
bronze and above groups so this is not necessary.
2. This will now also exclude `expired?` groups which we shouldn't have
been indexing to begin with. The assumption is that the `expired?` check
will also ensure we skip expired trials. These expired trials also have
a `trial_ends_on` date but based on looking at some data I think we can
expect this will be the same or less than `end_date`. The expired trials
that are in the index will be cleaned up in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/38551
3. Don't check `saved_changes.key?('hosted_plan_id')` anymore since we
don't really care why the plan is changing. As long as we use
`safe_find_or_create_by!` we'll be fine to run this multiple times.