Add cron worker to clean up expired subscriptions from Elasticsearch
Since paid subscriptions and trials will automatically be added to the
index we want a way to ensure they are cleaned up if their subscription
expires at some point. Removing from the `ElasticsearchIndexedNamespace`
table will be sufficient to trigger them to be deleted from the index.

We should wait at least 7 days after they are removed, however, to
ensure that we aren't wastefully deleting them if they end up upgrading
their plan in this window.