Add DynamicFields component
This adds a dynamic form field component, to be used in the [SAST
Configuration UI][epic].

So far, this component is only able to render a single type of form
field: a text input (via the [`FormInput`][input] component). Future
iterations will add support for more types.

This component is driven by an array of configuration entities (in fact,
a `SastCiConfigurationEntity` from GraphQL), each representing
a configurable variable when setting up SAST scanning for a project. In
a follow-up MR, this array will be provided by the GraphQL query
`Project.sastCiConfiguration`.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/231370.

[epic]: https://gitlab.com/groups/gitlab-org/-/epics/3659
[input]: https://gitlab.com/gitlab-org/gitlab/-/issues/225224