Update gRPC v1.30.2 and google-protobuf to 3.12.4
This will help unblock Ruby 2.7 support. We previously had to revert
these due to CentOS 6 build issues, but with
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/4466 we
are able to work around the compiler limitations.

The upgrade in google-protobuf from v3.8 to v3.9 had a breaking change:
https://github.com/protocolbuffers/protobuf/pull/6166. As
https://apidock.com/rails/Object/as_json mentions, Rails' `as_json` uses
`to_hash`, which no longer exists. Now we need to coerce values to
`to_h` to retain the original behavior.