Add Sidekiq worker context for project imports
When an import job is created on project creation, it doesn't have
project context metadata in our Sidekiq logs because the controller
doesn't know the project name yet. This fixes that.

It's not clear why the call to `#import_schedule` isn't guarded by
`@project.persisted?` in the same way as `#after_create_actions`, but
this preserves the existing behaviour.