Add guideline about resetting schema cache in migration docs
If using ActiveRecord, there may be cases where the schema is changed
in a previous migration and is cached, resulting in weird behaviors if
using the column that was altered but not refreshed.

As per https://github.com/rails/rails/pull/24300#issuecomment-201321397,
reset_column_information should be used when using models in migrations.