Populate SAST Configuration via GraphQL
Part of the [SAST Configuration UI][epic] feature, this adds a GraphQL
query to the SAST Configuration UI application, and feeds a transformed
copy of the response to the [`DynamicFields`][1] component in order to
render the form fields.

The usage of `DynamicFields` in this change is not ideal, but is only
written this way for now to demonstrate the data flow. This will be
improved in a related MR soon. Specifically, what's not ideal:

 - It currently mutates the fetched GraphQL response, rather than
   working with a copy of it.
 - It's not contained within a `form` element, and there's no way (yet)
   to _submit_ the form data.

Other changes include:

 - Replace Bootstrap utility class with one from GitLab UI.
 - Extract helper for populating data for the frontend.
 - Change subheading text to match latest design.
 - Move documentation link into subheading.
 - Add error state for app component.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/231372.

[epic]: https://gitlab.com/groups/gitlab-org/-/epics/3659
[1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/38162