Clean up orphaned LFS file references during GC
At the moment, users have no way to clean up orphaned LFS objects in
their repositories. Ideally, this would happen automatically, but in
the meantime, making it part of project housekeeping - which can be
manually triggered, and is also used during project cleanup, which is
often used as part of the "remove LFS from this repository" flow -
solves a pressing need.