Remove unused ml_IN locale file
This language is not being targeted in Crowdin or in the
GitLab::I18n::AVAILABLE_LANGUAGES list. Because of this it is not
receiving updates to strings from Crowdin and therefore is unable to
pass updated linting rules.

Remove this file so that we can pass the HTML linter when removing all
the strings from the todolist.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/228846