Don't show subscribable banner to auto_renew:nil
This is a bugfix. While 1576 has already been closed, we were informed
that we were displaying the banner to customers who's subscription was
auto_renew:nil. The problem ended up being that we were coercing the
value from nil to false while doing the check.

I've updated the code to use the attribute directly. I've also updated a
few of the method names to reflect that the value can be nil.

part of: https://gitlab.com/gitlab-org/growth/product/-/issues/1576