Only update pages config if pages are deployed
The pages config is a file inside the path on the shared NFS. The
directory into which this file is written is created when pages are
deployed and the artifacts are extracted.

If this path has been removed, or hasn't been created yet, we don't
need to update configuration.

Updating the configuration is also a step in the deployment of
pages. So the first deploy, the configuration would be written.