Drop minimum required Git version to v2.24.0 again
In order to be able to make use of the reference-transaction hook in
Gitaly, the minimum required Git version was bumped to v2.28.0. But due
to a bug in the hook's implementation in git-core it cannot yet be used
reliably, and thus the new requirement doesn't make a lot of sense yet.
So let's remove the requirement and instead make v2.28.0 the recommended
version.

Note that previous requirements were inconsistent. Some parts noted that
Git v2.24.0 is required starting with GitLab 13.1, while other parts
said Git v2.25.0 is required. To avoid a retroactive bump, this commit
adjusts documentation to v2.24.0.