Remove not-null on type column in audit_events
We are in the process of moving away from STI model in `AuditEvent` and
dropping `type` column. There is only one type (e.g. `SecurityEvent`)
and this redundant column wastes precious storage.

Rollback is an no-op since there would be null values added to the table
after the application stop writing to this field. We will need to create
separate migration to rectify these null values if there is a need to go
back to the original schema.