Update documentatiton on how to enable strong consistency
Starting with Gitaly 13.3, reference transactions are enabled by default
with a primary-wins strategy. While Gitalies will now vote on reference
updates by default, this still doesn't enable strong consistency as the
primary will always succeed no matter what secondaries vote for. As a
result, an admin now only has to change the voting strategy to not use
the primary-wins strategy in order to enable strong consistency.

Adjust the documentation to match this.