Geo - Generic Node Status Bars
This change removes the need for a hard coded
list of replicable types in the Geo Frontend
in regards to Node Status Bars.

Instead the frontend recieves the list and formats the
data from the API as it needs to match the existing
patterns.

This change also required some small backend
tweaks.