Make CreatePipelineService to run in dry-run mode
This allows a pipeline creation to fully run while skipping
the persistence steps. In the end it returns a non-persisted
pipeline with all its errors and warnings.

This feature is useful to allow CI Lint to use the actual
pipeline creation and display all errors rather than using
only YamlProcessor.