Use history icon on recent search filter tab only on mobile
The icon was replaced with the text to avoid ambiguity of the icon at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/18468, but this
partially reverts the original implementation with the icon only for
mobile/tablet screens to enlarge the input area of search in the
filter bar.

This MR does not include improving the width and/or the misalignment
of the search bar.

The original implementation of the icon only:
https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/10324

Signed-off-by: Takuya Noguchi <takninnovationresearch@gmail.com>