Log more lines for exceptions
Gitlab::BacktraceCleaner excludes some 'boring' lines from exception
stack traces: initializers, middlewares, logging code, etc.

This makes sense for development usage, or where we want to trace the
origin of (say) a database query. But in logs - particularly production
logs - we want all of our application code in the stack trace, because
we might have an error in our middleware, or we might be in a particular
code path triggered by a middleware.