Set proper exit status when Merge Trains fail
When triggering a merge train for stable branches, Curl would exit with
a status of 0 in the event of a failure (e.g. a wrong token being used).
By using the -f flag, we tell Curl to exit with a non-zero exit code
instead. This in turn will result in build failures, making it more
obvious to the Delivery team that something is wrong.