Return message when personal access token creation fails in internal API
This is a follow-up to the discussion in
https://gitlab.com/gitlab-org/gitlab-shell/-/merge_requests/397
to move expiry date validation onto the API side. See:

https://gitlab.com/gitlab-org/gitlab-shell/-/merge_requests/397#note_391081491)

Before this change the internal api would return a 500 error to the user
when creation of the personal access token fails, for example, when a
maximum token lifetime is enforced.

After this change, the internal api will return a more meaningful error
message to the user.