Make highlighting limits stricter
Currently, whenever we highlight text via `Gitlab::Highlight` as
rich text, it takes a long time (~3s). It falls back to plain text
highlighting after 3 seoncds due to the limit set via
`TIMEOUT_FOREGROUND`.

We are reducing that to 1.5 seconds so we don't have to wait that
long before falling back to plain highlighting.

We are also reducing `MAXIMUM_TEXT_HIGHLIGHT_SIZE` to 512 KB so
we don't need to wait 1.5 seconds before falling back as well when
highlighting a huge file.