Include Redis in Workhorse-in-test-suite integration
We started using Workhorse in feature tests, but no `config.toml` was
provided. As a result, we missed a startup issue that could have been
caught by testing the config loading.

A small subset of Workhorse functionality also relies on Workhorse and
Rails sharing a Redis database. If we configure it here, we can enhance
the runner API testing in rails to ensure it behaves as expected.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/209245