Implement authentication for the new Generic Packages API
This builds the foundation for the new Generic Packages API endpoionts.
Authentication with personal access token and job token is supported.

It's using mock `GET ping` endpoint that will be replaced with a real
one in a following MR.

This new API is behind `generic_packages` feature flag.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/235490.