Simplify docs on enable Elasticsearch integration
1. Remove instructions for indexing small instances. It's not necessary
since we already have instructions for enabling/indexing which works
fine for small instances.
1. Remove the separation between "indexing via rake" or "via
administration UI" considering that the via Rake option is only
necessary/useful for very large instances
1. Remove the section describing the 1MiB file size limit because this
is actually configurable now and the configuration option is already
described at the top of this section as "Maximum file size indexed"
1. Reduce the extra link back to explaining backfilling data since this
is part of the setup process and the big green button we have today
which says "Index all projects" is much clearer than the previous user
experience.