Move Pages directory in Sidekiq when renaming a project
This is behind the async_pages_move_project_rename feature flag. When
it's enabled and a project is renamed, we do not move the Pages
directory in Puma - instead, we schedule a Sidekiq job to perform the
move.

We already weren't checking the return value of
`Gitlab::PagesTransfer#rename_project`, so this shouldn't introduce any
failure cases that weren't there already.