Always attempt retry of job trace read when file is missing
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40075, we added
a check for the write lock to determine whether to retry reading a
missing, archived build log. However, in production, it appears this
issue is still happening about 1000 times a day, so let's always retry
to see if this reduces that incident rate.

This reverts commit f09bbb5808d1d9a3b57822233124a6ee6b9b10c8.