Fix permission to modify project-level MR rules
This change fixes issue where modification of project-level MR rules is
locked due to the instance-level settings. The permission to edit
project-level MR rules will no longer be controlled by instance-level
setting until MR rules are introduced at instance-level in upcoming
release.

The label of the instance-level setting is also updated to be on-par
with the one of project-level.