Utilize GitLab UI icon component: Web IDE
This replaces all the usages of ~/vue_shared/components/icon.vue with
the corresponding GitLab UI component within the Web IDE

In a previous MR [0] we already replaced the internal implementation
with the GitLab UI component. This should allow us to migrate all usages
without any test failures relating to the differences in their
implementation.

The actual deletion of the file will be done in a follow-up in case we
have missed an instance or someone introduces a new one in the
meanwhile.

[0]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/38778