Put Pages moves from namespace renames in Sidekiq
When we rename a namespace, any projects in that namespace with Pages
enabled have to have their directories moved on the filesystem to match.
We want all operations that touch Pages disk paths to happen in Sidekiq,
not Puma; this puts that in Sidekiq behind the
`async_pages_move_namespace_rename` feature flag.