Utilize GitLab UI icon component: pipelines to vue_merge_request_widget
This replaces all the usages of ~/vue_shared/components/icon.vue with
the corresponding GitLab UI component for all components from pipelines
to vue_merge_request_widget (alphabetically).

In a previous MR [0] we already replaced the internal implementation
with the GitLab UI component. This should allow us to migrate all usages
without any test failures relating to the differences in their
implementation.

The actual deletion of the file will be done in a follow-up in case we
have missed an instance or someone introduces a new one in the
meanwhile.

[0]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/38778