Put Pages moves from namespace transfers in Sidekiq
This is the same as the previous commit, but when a namespace changes
its parent namespace via a transfer. (The previous commit is for a
namespace changing its own path directly.) We're moving Pages disk
operations from Puma to Sidekiq, and here that's behind the
`async_pages_move_namespace_transfer` feature flag.