Redirect chatops requests to team channels
The production channel is for production incidents and operations, and
the infrastructure team are primarily responsible for it. Messages
asking to be added to chatops usually come to the attention of infra
before any other team. This leads to redundant attempts to re-validate
the legitimacy of the request, which could more easily be handled by a
fellow team member of the requester, in their channel.

SRE's on-call configure the production channel to alert them for every
new message, so chatops requests here are quite noisy.