Do not load MR records when requesting counts
This change adds `first: 0` to the throughput analytics count query to
avoid loading MR records into memory. Due to a preloader issue, even if
only the count is requested, the first 100 records are going to be
loaded. By specifying 0, the query will be still executed, however it
will be evaluated as no-op.