Introduce `Security::Finding` model
This model will represent all the findings discovered on all pipelines.
Unlike `Vulnerabilities::Finding` this model stores only the subset of
finding data which will be used to optimize the performance of pipeline
security tab by identifying which report artifact(s) to download and
parse while generating the paginated list of findings.