Make Vault Secret auth method path configurable
Currently Vault Secrets are using partially hardcoded configuration
for the authentication method. While the Vault server URL and the
role name for JWT authentication method can be configured with the
variables, the authentication method path is hardoced to `jwt`.
This may limit the usability of our solution.

This change makes this value configurable with the `VAULT_AUTH_PATH`
variable (similar to how `VAULT_SERVER_URL` and `VAULT_AUTH_ROLE`
are being used already) and ensures that in case when the variable
is not defined by the user, it will fall-back to the `jwt` value
that we have hardcoded now.