Remove error swallowing from pages config update
Updating the pages config would swallow all errors before. When we
moved this to Sidekiq, we re-raised the error from the worker, so we
could make use of the Sidekiq retries to try and recover from the
failed update.

Since we've moved this to Sidekiq, and are not performing config
updates if pages aren't deployed. We can now get rid of this weird
error handling that would cause the problem being reported to Sentry
twice.