Add invite email experiment
This creates an experiment for a new invite email design that is
simpler for our users and only has one call to action instead of
multiple. In the first iteration, this is only for new users and
we're tracking how many more users accept the invitation with
the new design compared to the old one.