Fix exception handling when a concurrent backup fails
If a concurrent backup encounters an exception, we should allow the
current threads to run to completion instead of aborting the whole
backup. This was causing threads to leak outside of the tests in
https://gitlab.com/gitlab-org/gitlab/-/issues/241251 and causing the
query cache to be enabled.

This also reverts the quarantined spec in
f47a2f17d14d1b100dbed3f176e55ec635268ca6.

Closes https://gitlab.com/gitlab-org/gitlab/-/merge_requests/40451