Split a Review Apps CI rule into two dedicated rules
In-order to prevent `review-performance` from running, a condition was
set in `.review:rules:mr-and-schedule-auto-if-frontend-manual-otherwise`
in `.gitlab/ci/rules.gitlab-ci.yml` file, but this also prevented
`review-deploy` from running as `review-deploy` also extended
`.review:rules:mr-and-schedule-auto-if-frontend-manual-otherwise`.

Since the `dast` jobs are set to `needs: ["review-deploy"]`, the CI
config was invalid. Unfortunately, when for scheduled pipeline, there's
no error logged or raised.

Signed-off-by: Rémy Coutable <remy@rymai.me>