Create Todos for designs
Allowing Projects::TodosController to create Todos for Designs.

This involved a small change to FinderMethods to refer to an objects
`#to_ability_name` method if defined, as due to its namespacing,
the existing deriving of from `DesignManagement::Design.model_name`
would not work. FinderMethods fallsback to its old deriving method for
when an object does not define `#to_ability_name`.

https://gitlab.com/gitlab-org/gitlab/-/issues/198439