Fix race condition in httprs test
The `So(server, ShouldNotBeNil)` call in the httprs test attempts to
print a string representation of the `server`, but this is racy because
the reconnection state is initialized in a
Goroutine. `httptest.NewServer()` always returns a valid value, so we
can drop this test.

Relates to https://gitlab.com/gitlab-org/gitlab-workhorse/-/issues/285